Unpack list of tuples
