+++ src/Analysis/Ast/Impl/Analyzer/Handlers/SequenceExpressionHandler.cs
@@ -30,6 +30,13 @@ namespace Microsoft.Python.Analysis.Analyzer.Handlers {
 
         internal static void Assign(SequenceExpression seq, IMember value, ExpressionEval eval) {
             var typeEnum = new ValueEnumerator(value, eval.UnknownType, eval.Module);
+            // Fetch actual tuple from the list for assignment item by item
+            // rather than assigning tuples from the list to each item.
+            // x = [('abc', 42, True), ('abc', 23, False)]
+            // for some_str, some_int, some_bool in x:
+            if (seq is TupleExpression && typeEnum.Peek is IPythonCollection coll) {
+                typeEnum = new ValueEnumerator(coll, eval.UnknownType, eval.Module);
+            }
             Assign(seq, typeEnum, eval);
         }
