+++ src/Analysis/Ast/Test/ValuesTests.cs
@@ -259,23 +259,5 @@ def g():
             analysis.Should().HaveFunction("g")
                 .Which.Should().HaveVariable("e2").OfType("MyException");
         }
-
-
-        [TestMethod, Priority(0)]
-        public async Task Enumerate() {
-            const string code = @"
-a = ['a', 'b', 'c']
-
-a1 = True
-a2 = False
-
-for i, x in enumerate(a):
-    a1 = i
-    a2 = x
-";
-            var analysis = await GetAnalysisAsync(code);
-            analysis.Should().HaveVariable("a1").OfType(BuiltinTypeId.Int)
-                .And.HaveVariable("a2").OfType(BuiltinTypeId.Str);
-        }
     }
 }
